DOCX textstyling - lists: numbering-id 0 misused
The numbering definitions of lists are stored within the docx file in
the file numbering.xml. If this file doesn’t exists or doesn‘t contain
any definition, then new definitions are created by xdocreport on the
fly starting by „numId“ zero. But according to the specification (Chap.
17.9.18) a numId of zero stands for a deleted numbering. So the first
list (numId 0) is not a list anymore. The other lists seems to get
messed up too. This commit changes this behavour and uses numIds
starting at 1.