Avoid blank pages at the end of sections.
This helps to avoid blank pages at the end if sections by ignoring pages
that only have blank space. Some of the new lines have been added by
XWPFDocumentVisitor.isAddNewLine returning true even when the page or
table is the last one in the section. While it would be possible to fix
XWPFDocumentVisitor.isAddNewLine not to add these, this commit also
solves the issue if the docx had an extra new lines at the end of the
page and the conversion to pdf had more space than the original docx.